🚚 Rename: AuthHeadline 페이지 컴포넌트에 작성 및 파일 이동
