+++ src/app/(public)/(auth)/ui/SpeechBubble.tsx
