+++ src/features/auth/ui/LoginClient.tsx
@@ -7,7 +7,6 @@ import { loginUser } from '@/features/auth/apis/auth';
 import { AUTH_API_MESSAGE } from '@/features/auth/constants/authMessage';
 import AuthFooter from '@/features/auth/ui/AuthFooter';
 import AuthForm from '@/features/auth/ui/AuthForm';
-import AuthHeadline from '@/features/auth/ui/AuthHeadline';
 import { LoginFormValues, loginSchema } from '@/features/auth/utils/schema';
 import { ApiError } from '@/shared/apis/apiError';
 import { useUserStore } from '@/shared/stores/userStore';
@@ -87,7 +86,6 @@ export default function LoginClient() {
 
   return (
     <>
-      <AuthHeadline />
       <div className="mt-17 md:mt-22">
         <AuthForm onSubmit={handleSubmit(handleLogin)}>
           <FormField label="이메일">
